chore: clean up types setup and migrate generic types
