+++ packages/client/src/clients/preferences/interfaces.ts
@@ -1,12 +1,4 @@
-// Channel types supported in Knock
-// TODO: it would be great to pull this in from an external location
-export type ChannelType =
-  | "email"
-  | "in_app_feed"
-  | "sms"
-  | "push"
-  | "chat"
-  | "http";
+import { ChannelType } from "@knocklabs/types";
 
 export type ChannelTypePreferences = {
   [K in ChannelType]?: boolean;
