--- packages/types/src/index.ts
@@ -1,3 +1,0 @@
-export type Sample = {
-  what: string
-}
